add user dependency for resquests

+++ todo/routers/todos.py
@@ -31,12 +31,17 @@ class Todo_request(BaseModel):
 
 
 @router.get("/",status_code=status.HTTP_200_OK)
-async def read_all(db : db_dependency):   
-    return db.query(Todos).all()
+async def read_all(user: user_dependency , db : db_dependency):  
+    if user is None :
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="user not authorised")
+    return db.query(Todos).filter(Todos.owner_id==user.get('id')).all()
 
 @router.get("/todo/{todo_id}",status_code=status.HTTP_200_OK)
-async def get_todo_by_id(db: db_dependency, todo_id : int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def get_todo_by_id(user : user_dependency,db: db_dependency, todo_id : int = Path(gt=0)):
+    if user is None :
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="user not authorised")
+    
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id==user.get('id')).first()
     if todo_model is not None:
         return todo_model
     else :
@@ -57,9 +62,12 @@ async def create_new_todo(user : user_dependency,db: db_dependency , todo : Todo
     return todo_model
 
 
-@router.put('/todos',status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db : db_dependency , todo : Todo_request, todo_id : int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id)
+@router.put('/todos/{todo_id}',status_code=status.HTTP_204_NO_CONTENT)
+async def update_todo(user : user_dependency,db : db_dependency , todo : Todo_request, todo_id : int = Path(gt=0)):
+    if user is None :
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="user not authorised")
+    
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404 , detail="not found")
     todo_model.title =  todo.title
@@ -74,14 +82,17 @@ async def update_todo(db : db_dependency , todo : Todo_request, todo_id : int = 
 
 
 @router.delete("/delete/{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db : db_dependency ,todo_id : int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user : user_dependency ,db : db_dependency ,todo_id : int = Path(gt=0)):
+    if user is None :
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="user not authorised")
+    
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id==user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404 , detail="not found")
     else :
-        db.query(Todos).filter(Todos.id == todo_id).delete()
+        db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id==user.get('id')).delete()
 
         db.commit()
-        db.refresh(todo_model)
+
